Added a function that samples doubles from the Uniform(0,1) distribution.
Most samplers from the Uniform(0,1) distribution simply choose values
which are integer multiples of 2**(-53).  Those are only a small fraction
of the doubles in the interval (0,1).  Another approach is to choose
the significand and the exponent separately and then combine the two
choices to construct a double.  This give significantly better
coverage of the possible values.

This sampler is used in the snapping mechanism which is frequently
suggested as a more robust replacement for the standard Laplace
mechanism.  The latter suffers from problems caused by standard
floating-point implementation details.  The former was designed
explicity to solve those problems.