Added a utility function to compute bits of a double
One of the atomic operations in the Snapping mechanism is to
found a double-precision floating-point number to the nearest
integer multiple of some power of two.  The original paper
recommends doing so by manipulating the bits of the double
directly.  Doing so requires some utility functions to first
extract and then manipulate those bits.  This is the first of
the required functions. I decided to dump the bits into a
unsigned 64 bit integer so that they would be easy to manipulate
with traditional bit shifting and masking operations.