Reimplemented the Sparse* Mechanisms
I found a paper (https://arxiv.org/pdf/1603.01699.pdf) that suggested
several improvements to the SparseVectorThreshold family of release
mechanisms.  I incorporated their suggestions into our implementation.

I also refactored the code a bit to do a better job of reusing code.
I added a parent class called SparseGeneric which implements the SVT
algorithm as described in the papaer.  I then made a tower of classes
that each inherit methods from their predecessor:

SparseGeneric -> SparseNumeric -> SparseIndicator -> AboveThreshold

SparseNumeric is just SparseGeneric with a different interface.  In
SparseGeneric, epsilon1, epsilon2, and epsilon3 are all specified
explicitly.  In SparseNumeric, the only the total epsilon is
specified explicitly.  The relative values of epsilon1, epsilon2,
and epsilon3 are specified by the parameters e2_weight and e3_weight
(e1_weight is implicitly always set to e1_weight = 1.0). This is in
keeping with the analysis in the paper cited above for the optimal
relative values of epsilon1 and epsilon2.

SparseIndicator simply sets `epsilon3=0.0` and passes that value along
with the arguments it received to the constructor for SparseNumeric.
Because we are only interested in the indices of the values that
exceed the specified threshold, SparseIndicator.release() only
returns the indices of the values that exceed the specified threshold.

AboveThreshold simply sets `cutoff=1` and passes that value along
with teh arguments it received to the constructor for SparseIndicator.
Because we are only interested in the index of the first element that
exceeds the specified threshold, AboveThreshold.release() returns that
value as in integer rather than a numpy array of length 1.