Added goodness of fit tests for all samplers
Added a test that tests the null hypothesis that our samplers
generate samples from the same distribution as the samplers
provided by `scipy.stats` against the alternative hypothesis
that they don't. We use a two-sided Kolmogorov-Smirnov test
to do the work via the command `scipy.stats.ks_2samp`

In the case of two_sided_geometric we compare our samples to
the difference of two geometric random samples as generated
by `scipy.stats.geom`.