Replaced sign/multiply fp_laplace with biased selection fp_laplace.
I replaced the simple method of generating a two-sided geometric
by generating a one-sided geometric and then independently generating
a sign bit.  This method overweights zero.

Instead, we generate a biased bit.  If that bit is 0, then we keep
our original geometric RV.  If that bit is 1, then we take the
bitwise complement of the original geometric RV. IF we choose the
biase correctly, then this gives us a true TwoSidedGeometric
distribution supported on exactly the set of 64-bit signed integers.

I also refactored the code that we use to generate bits drawn from
an exponential districution into a separate function.  Doing so makes
the code much easier to read, but also a bit slower.  I don't think
this structure works with the more efficient bit-coonsuming
generation strategy, so if we want to go back, we should probably
think hard about the code layout at that time.

Finally, I adjusted the range of the indices for the bits we
generate to exclude idx = 0.  This is because we really want to
generate a 63-bit unsigned integer which we use with the biased
selection bit to generate a 64-bit signed integer.