Precompute mix_bias
I replaced the bias for the most significant exponential bit
with the bias of the mix_bit.  This was free because we weren't
using that bias anyway.  This is because we changed from filling
all of the bits of the u64 to only filling the low 63 bits.  This
lets us use the biased selection to fill out an i64.

This version is much faster!