Modified Laplace mechanism to use fixed_point_laplace sampler.
Note: At this time fixed_point_laplace returns a floating point
value NOT a fixed point value.  This float is then combined with
the original values supplied to the mechanism.  This is NOT secure.
The next incremental step is to modify things so that
fixed_point_laplace returns a fixed point number which is added
to a fixed point approximation of the supplied values.  Then, this
perturbed version of the fixed point approximation can be re-cast
as a float and returned as the output of the release mechanism.