Code tweaks (#21)
This PR does a few things:

- fixes warning messages
- silences doc warnings
- refactors fixed point laplace

It also fixes a bug I introduced in my last PR. We now need to keep track of the random bits used in sample_exponential_bit and re-use them in the exact function. This necessitates passing the random bits around which also yields a 70% speedup. I changed the abstraction so that the sample_exponential_bit is now compare_exponential_bit and only requires rand_bits and bias as args.

* initial commit

* initial fix

* use all bits of bias and style changes

* use i64 for exponential bias throughout

* bias calcs in one vec

* fix rust warnings + silence rust doc warnings

* speedup fixed point laplace by 70%

* cleanup