Removed vestigial code.
I removed all of the code that we don't need any more.  Most of
this code is left over from earlier iterations of the
fixed_point_laplace sampler or uniform_double sampler.  There's
still work to do.  In particular, we can extract the code that
generates a fixed_point exponential random variate from the
fixed_point_laplace code and then use fixed_point_exponential
with precision=0 to generate geometric samples.

I also replaced the calls to the naive version of two_sided_geometric
in the Geometric mechanism with a call to fixed_point_laplace
with precision = 0. This rendered even more code surplus to
requirements so I went ahead and excised that code as well.