Implemented the "effective epsilon" trick.
Some of the literature describes all release mechanisms as if the
sensitivity of the underlying query is 1.0.  They then state that
these mechanisms privide epsilon * sensitivity - DP.  This has
the advantage of simplifiying the delicate and expensive arbitrary-
precision floating-point computations in the rust code.

As a part of this, I also modified the class strucutre a little bit.
Because the GeometricMechanism operates on integer-valued data and
adds integer-valued noise, we don't need to adjust the sensitivity
to account for floating-point to fixed-point conversion errors like
we do for the LaplaceMechanism.  As such, it made sense to change
the parent class of GemetricMechanism to the base ReleaseMechanism
class instead of LaplaceMechanism.  This simplified the constructor
for GeometricMechanism now that we're doing the effective epsilon
trick.