Changed the interface for hist.get_query_vector
The old version didn't let you specify an arbitrary list of
key/value pairs that a query should "hit".  To make that
possible, I changed the expected input from a dict of
key/value pairs to list of dicts of key value pairs.

It's a bit ugly, but necessary to capture the full range
of possible queries.